cloud-functions/sharepoint-direct-sync/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
pandas==2.*
requests==2.*
openpyxl==3.*
pyarrow==14.*
